code clean and set up
