+++ package.json
@@ -28,6 +28,7 @@
     "react-native-animatable": "^1.4.0",
     "react-native-audio-record": "^0.2.2",
     "react-native-audio-recorder-player": "^3.6.11",
+    "react-native-crypto-js": "^1.0.0",
     "react-native-fs": "^2.20.0",
     "react-native-gesture-handler": "^2.19.0",
     "react-native-linear-gradient": "^2.8.3",
