+++ src/navigation/index.js
@@ -1,4 +1,5 @@
-import LoginForm from '../Screens/Authentication/Login';
+import EmailForm from '../Screens/Authentication/emailLogin/email';
+import PasswordForm from '../Screens/Authentication/emailLogin/password';
 import SignUpForm from '../Screens/Authentication/SignUp';
 import RegisterAudioScreen from '../components/RegisterAudioPage';
 import Home from '../Screens/Home';
@@ -13,7 +14,8 @@ function Navigation() {
     <Stack.Navigator
       screenOptions={{headerShown: false}}
       initialRouteName="Home">
-      <Stack.Screen name="LoginForm" component={LoginForm} />
+      <Stack.Screen name="EmailForm" component={EmailForm} />
+      <Stack.Screen name="PasswordForm" component={PasswordForm} />
       <Stack.Screen name="SignUpForm" component={SignUpForm} />
       <Stack.Screen name="Home" component={Home} />
       <Stack.Screen name="RegisterAudioScreen" component={RegisterAudioScreen} />
